SRVLGIC-1079: Split kie-tools e2e tests execution into separate workflows (#405)

+++ packages/sonataflow-operator/internal/controller/profiles/preview/states_preview.go
@@ -242,5 +242,5 @@ func (h *deployWithBuildWorkflowState) isWorkflowChanged(workflow *operatorapi.S
 	if err != nil {
 		return false, err
 	}
-	return actualCRC != workflow.Status.FlowCRC, nil
+	return int64(actualCRC) != workflow.Status.FlowCRC, nil
 }
